Add files via upload
https://www.kaggle.com/datasets/emmanueldjegou/sales-dataset/data
This is where I pulled the data from. This is the first time I sought out data to practice on with the assistance of a module or pulled data from an internal source. I did some mild cleaning (uniformed the text format), created a pivot table and chart, and did a basic statistical analysis. I did review the data for insights. I might add a comment to this upload. I didn't do it now because my brain says it's time for bed, and it is in need of rest. Goodnight!
 Dang Nabit, I will get the gist of all of this. Brain fog really has done a number on my ability to think and learn.